Delete EnvironmentSettings completely
